19a, Untested! Power save, 2/10 upload when cloudy
If the battery mA reading is -50 or lower, then go into "night" mode for minute 2-9 of every 10 minutes. Turns on for 2 minutes, because it can take until the next minute for the U to be ready to transmit. No data saving yet, but the framework is beginning to come together with wxCache.ino. Cache is not actually used yet.