21r, basic camera functionality
Turns cameras on when voltage is high enough, turns them off when running negative too long or voltage gets too low.
Beginnings of "snapshot request" functionality, but it doesn't actually work yet. The auto-off is so aggressive, we can't keep the cameras on long enough to get manual early-morning or late-evening shots. Trying to add the ability to take a quick snap.